update to automodel and autotokenizer
