+++ flambe/nlp/transformers/__init__.py
@@ -1,14 +1,8 @@
-from flambe.nlp.transformers.bert import BertTextField, BertEmbedder
-from flambe.nlp.transformers.roberta import RobertaTextField, RobertaEmbedder
-from flambe.nlp.transformers.gpt import GPTTextField, GPTEmbedder, GPT2TextField, GPT2Embedder
-from flambe.nlp.transformers.xlm import XLMTextField, XLMEmbedder
-from flambe.nlp.transformers.xlnet import XLNetTextField, XLNetEmbedder
-from flambe.nlp.transformers.xl import TransfoXLTextField, TransfoXLEmbedder
+from flambe.nlp.transformers.field import PretrainedTransformerField
+from flambe.nlp.transformers.model import PretrainedTransformerEmbedder
 from flambe.nlp.transformers.optim import AdamW, ConstantLRSchedule
 from flambe.nlp.transformers.optim import WarmupConstantSchedule, WarmupLinearSchedule
 
 
-__all__ = ['BertTextField', 'BertEmbedder', 'RobertaTextField', 'RobertaEmbedder', 'GPTTextField',
-           'GPTEmbedder', 'GPT2TextField', 'GPT2Embedder', 'XLMTextField', 'XLMEmbedder',
-           'XLNetTextField', 'XLNetEmbedder', 'TransfoXLTextField', 'TransfoXLEmbedder',
+__all__ = ['PretrainedTransformerField', 'PretrainedTransformerEmbedder',
            'AdamW', 'ConstantLRSchedule', 'WarmupConstantSchedule', 'WarmupLinearSchedule']
