--- flambe/nlp/transformers/xlm.py
@@ -1,48 +1,0 @@
-"""
-Intergation of the transformers xlm module.
-
-Note that these objects are only to be used to load
-pretrained models. The pytorch-transformers library
-wasn't designed to train these models from scratch.
-
-"""
-
-import transformers as pt
-
-from flambe.nlp.transformers.utils import TransformerTextField, TransformerEmbedder
-
-
-class XLMTextField(TransformerTextField):
-    """Integrate the transformers XLMTokenizer.
-
-    Currently available aliases:
-        . `xlm-mlm-en-2048`
-        . `xlm-mlm-ende-1024`
-        . `xlm-mlm-enfr-1024`
-        . `xlm-mlm-enro-1024`
-        . `xlm-mlm-tlm-xnli15-1024`
-        . `xlm-mlm-xnli15-1024`
-        . `xlm-clm-enfr-1024`
-        . `xlm-clm-ende-1024`
-
-    """
-
-    _cls = pt.XLMTokenizer
-
-
-class XLMEmbedder(TransformerEmbedder):
-    """Integrate the transformers XLMModel.
-
-    Currently available aliases:
-        . `xlm-mlm-en-2048`
-        . `xlm-mlm-ende-1024`
-        . `xlm-mlm-enfr-1024`
-        . `xlm-mlm-enro-1024`
-        . `xlm-mlm-tlm-xnli15-1024`
-        . `xlm-mlm-xnli15-1024`
-        . `xlm-clm-enfr-1024`
-        . `xlm-clm-ende-1024`
-
-    """
-
-    _cls = pt.XLMModel
